requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
crispy-bootstrap5==0.6
cryptography==37.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
dj-email-url==1.0.5
Django==4.0.5
django-allauth==0.51.0
django-cache-url==3.4.2
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
django-environ==0.9.0
django-jalali-date==1.0.1
django-js-asset==2.0.0
django-rosetta==0.9.8
environs==9.5.0
idna==3.3
jdatetime==4.1.0
marshmallow==3.16.0
oauthlib==3.2.0
packaging==21.3
polib==1.1.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-dotenv==0.20.0
python3-openid==3.2.0
requests==2.28.0
requests-oauthlib==1.3.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
